separated Edge and Vertex Classes into separate java class files.  Tested by running the code.
