+++ src/main/java/com/down2thewire/WeightedGraph.java
@@ -7,245 +7,6 @@ public class WeightedGraph {
     LinkedList<Edge> edgeList;
 
 
-    static class Vertex {
-        Location location;
-        String vertexName;
-        // 0-walk, 1-drive, 2-rideshare, 3-carRental, 4-bicycle, 5-scooter, 6-transit, 7-bus, 8-airplane, 9-unused
-        /*
-         * Walking is allowed in most places with stops. It is not allowed on the Interstate
-         * Changing modes to/from driving requires a parking lot, and the cost of parking should be added to the final driven leg
-         * Rideshare is allowed almost anywhere. In some places like the airport, there are specific pickup areas.
-         * Car Rental locations - where you can rent a car, for instance, the airport car rental lots
-         * Bicycle is allowed in most places with stops, but changing to/from bike means bike rental or locking rack.
-         * Biking is not allowed on the Interstate
-         * Scooter Rental Locations - where you can rent a scooter.
-         */
-
-        private boolean[] modes = {false, false, false, false, false, false, false, false, false, false};
-        private double tentativeDistance;
-
-        public Vertex(Location location, String uniqueNameId) {
-            this.location = location;
-            this.vertexName = uniqueNameId;
-            this.tentativeDistance = Double.POSITIVE_INFINITY;
-        }
-        public Vertex(){}
-
-        public Double getLongitude() {
-            return this.location.getLongitude();
-        }
-
-        public Double getLatitude() {
-            return this.location.getLatitude();
-        }
-
-        public void autoVertexName() {
-            // Get Location Name from GoogleLocation API
-            // ToDo
-        }
-
-        public void setWalk(boolean tf) {
-            modes[0] = tf;
-        }
-
-        public void setCarPark(boolean tf) {
-            modes[1] = tf;
-        }
-
-        public void setRideShare(boolean tf) {
-            modes[2] = tf;
-        }
-
-        public void setCarRental(boolean tf) {
-            modes[3] = tf;
-        }
-
-        public void setBikeRack(boolean tf) {
-            modes[4] = tf;
-        }
-
-        public void setScooterRental(boolean tf) {
-            modes[5] = tf;
-        }
-
-        public void setTransit(boolean tf) {
-            modes[6] = tf;
-        }
-
-        public void setBusStop(boolean tf) {
-            modes[7] = tf;
-        }
-
-        public boolean getWalk() {
-            return modes[0];
-        }
-
-        public boolean getCarPark() {
-            return modes[1];
-        }
-
-        public boolean getRideShare() {
-            return modes[2];
-        }
-
-        public boolean getCarRental() {
-            return modes[3];
-        }
-
-        public boolean getBikeRack() {
-            return modes[4];
-        }
-
-        public boolean getScooterRental() {
-            return modes[5];
-        }
-
-        public boolean getTransit() {
-            return modes[6];
-        }
-
-        public boolean getBusStop() {
-            return modes[7];
-        }
-
-        public double getTentativeDistance() {
-            return tentativeDistance;
-        }
-
-        public void setTentativeDistance(double tentativeDistance) {
-            this.tentativeDistance = tentativeDistance;
-        }
-
-        public List<Edge> getOutgoingEdges(List<Edge> edgeList) {
-            List<Edge> outgoingEdges = new ArrayList<>();
-            for (Edge edge : edgeList) {
-                if (edge.start.isMatch(this)) {
-                    outgoingEdges.add(edge);
-                }
-            }
-            return outgoingEdges;
-        }
-
-
-
-        public boolean isMatch(Vertex tempVer) {
-            // match rounded to 4 decimal places - about 10 meters, so 20-meter square box around location
-            if (this.location.isMatch(tempVer.location)) {
-                return Boolean.TRUE;
-            }
-            return Boolean.FALSE;
-        }
-    }
-
-
-    static class Edge {
-        Vertex start;
-        Vertex end;
-        String mode;
-        Integer duration;
-        Double cost;
-        Integer distance;
-
-        public Edge(Vertex start, Vertex end, String mode, Integer duration, Double cost, Integer distance) {
-            this.start = start;
-            this.end = end;
-            this.mode = mode;
-            this.duration = duration;
-            this.distance = distance;
-            this.cost = cost;
-            if (cost == 0.00d) {
-                this.cost = estimateCost(this);
-            }
-        }
-        public Vertex getStart() {
-            return start;
-        }
-
-        public String getMode() {
-            return mode;
-        }
-
-        public double getCost() {
-            return cost;
-        }
-        public Integer getDuration() {
-            return duration;
-        }
-        public static double estimateCost(Edge e){
-            switch(e.mode){
-                case "walking":
-                case "bicycling":
-                    return 0.00d;
-                case "driving":
-                    return (.90d * e.distance);
-                case "rideshare":
-                    return (10.0d + 1.60d * e.distance);
-                case "carrental":
-                    return (90.0d + .20 * e.distance);
-                case "scooter":
-                    return 1.0d + .39d * e.duration;
-                case "transit":
-                case "bus":
-                    return 2.50d;
-                default:
-                    return 0.00d;
-            }
-        }
-
-        static public int getMode (String mode) {
-            switch(mode) {
-                case "walking":
-                    return 0;
-                case "driving":
-                    return 1;
-                case "rideshare":
-                    return 2;
-                case "carrental":
-                    return 3;
-                case "bicycling":
-                    return 4;
-                case "scooter":
-                    return 5;
-                case "transit":
-                    return 6;
-                case "bus":
-                    return 7;
-                case "airplane":
-                    return 8;
-                default:
-                    return -1;
-            }
-        }
-        static public String getMode (int i){
-            switch(i) {
-                case 0:
-                    return "walking";
-                case 1:
-                    return "driving";
-                case 2:
-                    return "rideshare";
-                case 3:
-                    return "carrental";
-                case 4:
-                    return "bicycling";
-                case 5:
-                    return "scooter";
-                case 6:
-                    return "transit";
-                case 7:
-                    return "bus";
-                case 8:
-                    return "airplane";
-                default:
-                    return " ";
-            }
-        }
-        public Vertex getEnd() {
-            return end;
-        }
-    }
-
-
     public Edge getEdge(Vertex start, Vertex end) {
         for (Edge edge : edgeList) {
             if (edge.getStart().equals(start) && edge.getEnd().equals(end)) {
@@ -382,15 +143,17 @@ public class WeightedGraph {
         }
     }
 
-    public void cloneWgToListLevel(WeightedGraph g) {
+    public WeightedGraph cloneOfWgAndLists() {
+        // Vertices and Edges are NOT cloned
         WeightedGraph cloneWG = new WeightedGraph();
         cloneWG.vertexList = new LinkedList<Vertex>();
-        for (Vertex vertex : cloneWG.vertexList) {
+        for (Vertex vertex : this.vertexList) {
             cloneWG.addJustVertex(vertex);
         }
-        for (Edge edge : cloneWG.edgeList) {
+        for (Edge edge : this.edgeList) {
             cloneWG.addJustEdge(edge);
         }
+        return cloneWG;
     }
